LF-4078 - add alternate row color
