+++ packages/webapp/src/assets/colors.scss
@@ -65,8 +65,11 @@
   --modalPrimary:  white;
   --bgInputListTile: white;
   --mainBackground: #FAFCFB;
+  --tableAlternateRowBackground: #F7FBFF;
+
 
     // New design system colours
     --Colors-Primary-Primary-teal-900: #16423d;
     --Colors-Accent---singles-Brown-dark: #633700;
+
 }
