+++ packages/webapp/src/components/Table/Table.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2023 LiteFarm.org
+ *  Copyright 2024 LiteFarm.org
  *  This file is part of LiteFarm.
  *
  *  LiteFarm is free software: you can redistribute it and/or modify
@@ -12,60 +12,11 @@
  *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
-import type { ReactNode } from 'react';
+
+import React from 'react';
 import ReactTable from './index';
-import type { ColumnInstance } from 'react-table';
 import MuiTable from './v2';
-import { TableType } from './constants';
-import { ReactComponentLike } from 'prop-types';
-
-enum Alignment {
-  LEFT = 'left',
-  RIGHT = 'right',
-}
-
-type CallbackFn = (props: any) => any;
-
-type ReactTableProps = {
-  columns: ColumnInstance[];
-  data: Object[];
-  showPagination: boolean;
-  pageSizeOptions: number[];
-  defaultPageSize: number;
-  className: string;
-  getTdProps: Function;
-  sortByID: number;
-  minRows: number;
-  orderDesc: boolean;
-};
-
-type MuiTableColumn = {
-  id: string;
-  format: CallbackFn;
-  align: Alignment;
-  Footer: ReactNode;
-  columnProps: Object;
-  label: string;
-};
-
-type MuiTableProps = {
-  columns: Required<MuiTableColumn[]>;
-  data: Object[];
-  showPagination: boolean;
-  pageSizeOptions: number[];
-  minRows: number;
-  dense: boolean;
-  FooterCell: ReactComponentLike;
-  onClickMore: CallbackFn;
-  itemsToAddPerLoadMoreClick: number;
-  onRowClick: CallbackFn;
-  shouldFixTableLayout: boolean;
-  defaultOrderBy: string;
-};
-
-type KindComponentKVP = {
-  [kind in TableType]: (props: any) => ReactNode;
-};
+import { TableType, KindComponentKVP, ReactTableProps, MuiTableProps } from './types';
 
 const tableComponents: KindComponentKVP = {
   [TableType.V1]: (props: ReactTableProps) => <ReactTable {...props} />,
